security: transition to jwt header
