+++ src/api/types/query.ts
@@ -1,18 +1,18 @@
-import { z } from "zod";
+import {z} from "zod";
 
 /**
  * Request body schema for the Query tool/endpoint.
  */
 export const QueryRequestSchema = z.object({
-	query: z.string().min(1, "Query must be at least 1 character long"),
+    query: z.string().min(1, "Query must be at least 1 character long"),
 });
 
 /**
  * Response data schema for the Query tool/endpoint.
  */
 export const QueryResponseSchema = z.object({
-	response: z.string(), // 🔥 always required
-	metadata: z.record(z.any()).nullable(),
+    response: z.string(),
+    metadata: z.record(z.any()).nullable(),
 });
 
 export type QueryRequest = z.infer<typeof QueryRequestSchema>;
