+++ src/features/auth/utils/authHeaderUtils.ts
@@ -1,0 +1,5 @@
+export function getAuthHeader(accessToken?: string): { Authorization: string } | {} {
+    const jwt = accessToken;
+
+    return jwt ? {Authorization: `Bearer ${jwt}`} : {};
+}
