--- src/mcp/api/client/client.ts
@@ -1,9 +1,0 @@
-import { env } from "cloudflare:workers";
-import type { ApiClientConfig } from "@/mcp/api/client/config";
-import { createApiClient } from "./base";
-
-const KollektivAPIConfig: ApiClientConfig = {
-	baseUrl: env.API_BASE_URL,
-};
-
-export const api = createApiClient(KollektivAPIConfig);
