--- src/mcp/api/client/config.ts
@@ -1,10 +1,0 @@
-/**
- * API Client configuration
- */
-export interface ApiClientConfig {
-	baseUrl: string;
-	headers?: Record<string, string>;
-	timeout?: number;
-}
-
-export type HttpMethod = "GET" | "POST" | "PUT" | "PATCH" | "DELETE" | "HEAD" | "OPTIONS";
