+++ tests/mocks/index.ts
@@ -1,1 +1,2 @@
 export * from "./supabase-mocks";
+export * from "./hono-mocks";
