--- tests/unit/factories.ts
@@ -1,11 +1,0 @@
-import type { DocumentChunk } from "@/mcp/api/types/ragTasks";
-
-// DocumentChunk factory
-export const chunkFactory = (overrides: Partial<DocumentChunk> = {}): DocumentChunk => ({
-	documentFilename: "test_document.pdf",
-	title: "Document Title",
-	section: 1,
-	text: "This is the text of the document.",
-	relevanceScore: 0.9,
-	...overrides,
-});
